backend/lang_service/requirements.txt
psycopg2-binary
python-dotenv
langchain
langchain-openai
langchain_experimental
Flask
flask-cors
pyepsilla
sentence_transformers
pypdf
tiktoken
boto3
tweepy
faker